Added a simple innermost rewriter (as library libgsrewrite) and a rewrite tool (rewr) that uses it.
To use the rewriter, include the header file libgsrewrite.h and add
link option -lgsrewrite. The rewriter is initialised with a call
gsRewriteInit(ATermAppl Spec, int strat), where Spec should be a
specification containing the data equations and strat is one of the
available rewrite strategies (currently only GS_REWR_INNER). To
rewrite a term, one calls ATermAppl gsRewriteTerm(ATermAppl Term).

The rewr tool takes a specification (from the command line or stdin)
and rewrites the data in the processes. (Option -a can be used to
read a taf/baf file.)


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@270 21252fe5-60e8-0310-8ff0-c1ae1001594e