Added __APPLE__ to:
#if !(defined __USE_SVID || defined __USE_BSD || defined __USE_XOPEN_EXTENDED|| defined __APPLE__)

on line 17 of gslowlevel.

to avoid that two versions of strdup will be available on the MAC. This causes the compilation
of the toolset to fail. It is not clear to me whether this is the canonical way to detect
that the compiling computer is an Apple but it works.





git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@296 21252fe5-60e8-0310-8ff0-c1ae1001594e