Adapted error messages to gsErrorMsg, generally followed by stop() to terminate execution of the program.
Adapted several type errors, such that flags --cluster, --newstate and
others now seem to work with a number of tests that I ran. Problems
where often wrongly typed variables.

With the --nocluster option mcrl22lpe takes now too much time. This
is due to the combination of a large number of summands and an inefficient
way of removing superfluous delta summands.

Jan Friso



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@541 21252fe5-60e8-0310-8ff0-c1ae1001594e