Attempted to find a solution to find the correct name of boost libraries. In some cases a compiler suffix (e.g `-gcc') must be added. Some installations do not provide appropriate symlinks to versions of the libraries without such a suffix.
This revision changes the default to no prefix. I hope this solves the linking-to-boost
problems for most people at this point. Note that as a result of this change, 
a standard build and install of boost (1.33) from its source distribution will not
have a correct Makefile generated at the moment.

A suitable test in the configure script, to find whether and what suffix to use
for the boost libraries, will follow later.


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@586 21252fe5-60e8-0310-8ff0-c1ae1001594e