Substituted old Studio prototype with the new one.
The current setup is as follows. The current studio architecture consists of
four components. A project manager, a tool manager, a tool executor and lastly
the GUI.

The new prototype consists of a wxWidgets GUI build on top of a first version
of the project manager. The tool manager is still in early development and is
not included. And the tool executor has yet to be designed as well as build.
In its current state you should be able to get a picture about the
possibilities and limitations of this prototype. Please let me know.

There also exists documentation, that at the time consists of a basic
description of the architecture. I've received quite some comment from Aad
about an early version of this document and it is not yet processed.
Additionally the architecture has undergone some changes due to experience
during recent development.  This documentation is not included because it does
no longer reflect the current design of the architecture of the prototype. This
will be remedied soon.

From now on I will regularly commit changes.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@721 21252fe5-60e8-0310-8ff0-c1ae1001594e