Changes: - file renamings:   gsfunc     -> libstruct   gslowlevel -> liblowlevel - libprint_c is now compiled and linked using g++
From now on, the src directory does not contain any files with a gs prefix.
Also it does not contain any C files anymore. This means that it isn't
necessary to wrap source files in an 'extern "C"' declaration. More
importantly, this makes it easier to remove the C interfaces of the rewriter
and the pretty printer.


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@808 21252fe5-60e8-0310-8ff0-c1ae1001594e