Added transparency to the tool. Level of transparency can be set on the Settings tab.
Note that transparent objects are not always rendered correctly. This is a known
problem which is hard to solve, in the sense that it requires lots of code to be
rewritten and (when solved) probably decreases performance significantly. I'm
looking for ways to render objects correctly in more cases without having to
completely rewrite the application and decreasing performance too much.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@1173 21252fe5-60e8-0310-8ff0-c1ae1001594e