Changed checking for removable delta summands, to improve its efficiency. As the performance is far insufficient for Yaroslav's recent example (with multiactions of size 16 (sic)), two measures have been taken:
1) it is indicated how many new summands are added by the communication
   operator in verbose mode. The number of these summands is generally
   very substantial with large multiactions.
2) an option has been added (-g), (--no-deltaelm) to prevent the system
   from removing delta summands in the end result.

The reason for the problems are that removing delta summands is quadratic
in the number of summands. In Yaroslav's example there are appr. 43000
summands, which are the result of applying the communication operator
to 3 summands containing those rather sizable multiactions.

It is my impression that none of the above measures solve the problem in whatever
sense, but as it stands it seems the best I can do.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@1197 21252fe5-60e8-0310-8ff0-c1ae1001594e