Added a `trick' to the lineariser to avoid generating a large number of delta summands. When a communication on a multiaction is calculated, many new summands are generated. In particular, if the original summand has condition c, all new summands have conditions c && c1, c && c2,..., c && c_n. The conditions c1 || c2 || ... || c_n is a tautology.
When later an allow is applied most of these summands are
renamed to delta summands. This means there will be a large
number of delta summand with conditions c && c_i. It is not trivial
to remove them, because the conditions can be very complex.

However, and this is the trick, if an additional delta summand is added
with condition c when calculating the communication, then the lineariser
only has to check that c && c1  -> c to remove the delta summand
with the first condition. This works (reasonably) well.

For the example of Yaroslav, removing delta summands has become
rather straightforward now. The only resulting delta summand
has the shape

  true -> delta

which, I believe is an acceptable overhead.

Jan Friso



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@1203 21252fe5-60e8-0310-8ff0-c1ae1001594e