Implemented some improvements to the jittyc rewriter. For each function rewr_f(a,b,..) that returns the result of rewriting term 'f(a,b,..)', we now introduce functions rewr_f_00.., rewr_f_10.., rewr_f_01, rewr_f_11.. etc. to be able to keep track of normal forms. The bits added to the name indicate whether or not the i'th parameter is already in normal form or not. Besides this, also additional function symbols f_00, etc. are introduced to indicate the same information within terms.
For example, for function f with the rule f(x,y) -> g(h(x),y) a function
rewr_f_11 is generated to rewrite a term f(t,u) for which both t and u are
already in normal form. The application of the rule within rewr_f_11 would
result in a call to rewr_g_01(h_1(x),y). The rewriting of h_1(x) within
rewr_g_01 results in a call to rewr_h_1(x).

Note that this introduces a lot of extra functions and is therefore limited
to functions of arity smaller or equal to some N, which is currently set at
3.


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@1245 21252fe5-60e8-0310-8ff0-c1ae1001594e