transport:  - cleaned up a bit so boost::apply and boost::protect are not necessary anymore
sip:
 - modified messenger so that it uses conditions to block until a delivery
   event occurs instead of a barrier
 - added locking to make sure that no deadlock will occur in some rare
   situations involving multiple threads that are waiting for the same type of
   message to be delivered

A remaining issue is that due to multithreading it has become more probable
that delivery event handlers block to wait for something. This means that
during this time no messages can be processed (which is unacceptable). At the
moment this seems to be the only problem that prevents the squadt
initialisation process to succeed. If ran with a squadt enabled lpeinfo, then
initialisation will deadlock due to this. If lpeinfo is killed manually, then
initialisation finishes without taking down squadt with it...

A solution is almost implemented, but there are only so many hours in a day...


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@1352 21252fe5-60e8-0310-8ff0-c1ae1001594e