Increased the initial size of the array objectdata as an intermediate increase of this array leads to data loss, as an example of Yaroslav showed.
Essentially the problem occurs in statements of the
following form:

   objectdata[i].field=F(objectdata[i].field);

If objectdata is resized using realloc within F the
addresses objectdata[i] at the right hand side are
different from the one at the left hand side. The
compiler however uses the same addresses at the left
and the right hand side, leading to loss of data.

This problem has not really been solved, but
has been avoided by increasing the initial size
of the objectdata array.

The fundame


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@1515 21252fe5-60e8-0310-8ff0-c1ae1001594e