Repaired a bug found by Yaroslav where instead of the type of an action the argument of the action was used as subsequent sort. This error has been lingering for a year and ought to have been caught by the linear process property checker. Many generated linear processes are likely to have a term at some position where a sort is expected.
Removed a debug message used to see whether the connection
to squadt was working properly. The connection to squadt
is as yet still not functioning. Only a menu with message
options is shown.

Removed a mistake in the handling of arguments of the -p
flag. This error was introduced when encoding the squadt 
interface. This also should deal with the warnings of 
Bas.







git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@1619 21252fe5-60e8-0310-8ff0-c1ae1001594e