- Integrated lpeuntime with squadt
Note: I have as of now not tested the squadt integration, for I seem to have some trouble compiling squadt and others on various systems.
jwulp, could you have a look if the current source tree is compiling allright?

Note2: lpeuntime is complete for functionality, I have to see whether to give the dummy variable in the sum another name than "t" in order to prevent name collisions.


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@1910 21252fe5-60e8-0310-8ff0-c1ae1001594e