Added "Save Picture" functionality: the picture of the state space can now be easily saved as a bitmap file. The infamous screenshot-method is no longer needed.
An advantage over the screenshot-method is that the user can set the resolution
of the image that will be produced. Resolutions range from (1,1) to whatever
OpenGL supports, which is surely high enough to obtain high quality pictures.

Possible file formats are those that are supported by wxWidgets and include BMP,
PNG, JPEG, GIF, etc.

The functionality is accessible via the "File" menu or "Ctrl+P" shortcut.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@1979 21252fe5-60e8-0310-8ff0-c1ae1001594e