Modified:
  lts.cpp: Rewrote second pass of algorithm to be more readable.
  utils.h: Adpated treshold.
  cluster.*: Restructured slot resolvement into two phases:
    0) Place undecided states into slots.
    1) Distribute states over the available space of each slot.

Currently, the placement looks quite nice. The only problem still in state placements is that multiple states that are all centered are placed on top of each other. A possible solution for this is distributing these states evenly on a circle with radius tau, but I need to look into this.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@2482 21252fe5-60e8-0310-8ff0-c1ae1001594e