- Added Doxyfiles, HTML files, and the CSS stylesheet needed to generate the new   library reference website. Every library has its own Doxyfile that @includes a   master Doxyfile called "Doxymaster" containing common settings. - Added a make target to generate the website. Calling 'make doxy' from trunk/   will generate the website in trunk/build/doxy/. - Removed Doxyfile for old website.
The new website is available at http://www.mcrl2.org/libref/.

For now, I have considered every subdirectory of trunk/libraries/ to be a
library (and SIP as well), so each of these libraries now has its own Doxyfile
and appears as a separate menu item in the navigation bar of the website. This
might change in the near (though not as near as July 1) future.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@3113 21252fe5-60e8-0310-8ff0-c1ae1001594e