- Implemented attempt to improve state positioning. If too many states are in a   slot (a section of the cluster's rim), the states are now placed inside the   cluster towards its center, not just on the rim. This fixes problems where   states overlapped each other. It works pretty good on, for example,   dining7.fsm. However, in some state spaces there are still too many states in   a slot. I'm working on that. - Fixed a bug: after opening a new LTS, absolute state positions should be   updated. - Refactored parts of the state positioning code to allow for the improvements,   make it shorter and/or make it a bit more efficient in some places.
This implementation also supports positioning states on multiple rings inside a
cluster (not just the outermost rim). This was my first attempt to solve the
problems mentioned above, but it did not work as good as expected. The number of
rings is currently set to two (center and rim) which is equivalent to the old
situation. The code should be retained though, so I include it in this commit.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@3205 21252fe5-60e8-0310-8ff0-c1ae1001594e