Implemented a quantifier elimination algorithm based on instantiation of variables with constructors. This works the same for finite sorts as pbes2bes, but for infinite sorts, this procedure can terminate also, making it essentially more powerful.
Unfortunately, as it stands, many copies can be made of 
existing terms. E.g. a term forall x:Pos.phi(x)&&psi where
x does not occur in phi, is replaced by:

  forall x0:Nat, b:Bool. [(phi(@cDub(b,x0) && psi) && (phi(@c1) && psi)]

This in turn will be replaced by:

  forall x0:Nat.
     [(phi(@cDub(true,x0) && psi) && (phi(@c1) && psi)]&&[(phi(@cDub(false,x0) && psi) && (phi(@c1) && psi)]

The problem is the duplication of psi. I will look into this. If these problems
are resolved, I will make a library routine that gets a pbes expression with quantifiers
and tries to replace those.

Furthermore, there is a problem with the hashtables, which are incorrectly removed at the
end of pbes2bool. I will take this up with Wieger when he returns.

Jan Friso



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@3206 21252fe5-60e8-0310-8ff0-c1ae1001594e