Merge di_postpone branch as of r3262 to trunk. To summarize, this includes the   following changes: - The first part of data reconstruction is incorporated in the toolset. For   now data reconstruction is only used in the pretty printer. - Current data reconstruction does the following:   * Reconstruct implementation of binders (except the whr binder)   * Remove system defined functions for lambda expressions from the     data specification and map specification.   * Reconstruct numeric expressions.
This means that a specification in which binders occur are now much more intuitive to read when pretty printed.

Note that there is more work to be done, and this is by no means the final result. Nevertheless Aad and me think this is a great leap forward, and should therefore be included in the trunk.


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@3263 21252fe5-60e8-0310-8ff0-c1ae1001594e