- Removed some debug statements from constelm. - Added a workaround for undefined behavior of the ATerm Library in constelm.
In the code fragment below the bottom_of_stack variable doesn't mark the
bottom of the stack (Visual C++ 8, release mode). This results in undefined
behavior, which can be resolved by a call to spec1.protect(), as pointed out
by Muck.

N.B. Both the old an the new implementation of constelm suffer from this problem!

int main()
{
  aterm bottom_of_stack;
  aterm_init(bottom_of_stack);
  gsEnableConstructorFunctions();
  string infile(argv[1]);
  specification spec1;
  spec1.load(argv[1]);


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@3830 21252fe5-60e8-0310-8ff0-c1ae1001594e