Added parts to the body of default_expression(s) that for a sort s yields a default term. Currently, a term is only returned if it is a constant.
Furthermore, adapted pbes2bool, as it could not operator on PBESses
with implications and negations in it. Now the pbes is first normalized
before being processed. Also added a welltyped check.

Changed the message whether a pbes was valid into something more
comprehensible, because the current result was sometimes confusing.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@4139 21252fe5-60e8-0310-8ff0-c1ae1001594e