Fixed bug in the parser: constants of the form "-[1-9][0-9]*" were not parsed as integer constants but as negations of a positive number. Although this is still correct with respect to type checking, it resulted in problems when using negative integer constants in the left-hand sides of equations. For instance, consider the following data type specification:
map f: Int -> Bool;
eqn f(-1) = false;

When "f(-1)" was used somewhere else in the specification, it would not
rewrite to false.

Note that the issue is still present when using "f(- 1)" or "f(-p)",
where "p" is a variable of sort Pos, as a left-hand side of an equation.
This will be resolved when pattern matching is implemented.

This bug was found by Rink Springer.


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@4319 21252fe5-60e8-0310-8ff0-c1ae1001594e