Structure of the website has changed and the library reference pages now reflect these changes.
PHP functions have been defined in a single location for all common
elements on the website (e.g. the large mCRL2 logo bar and navigation
bar at the top). These are contained in
http://www.mcrl2.org/common/mcrl2.php and are now used by both the Wiki
site and the library reference pages.

Also, the library pages no longer use the Doxygen search engine, but now
use the same engine as I installed for the Wiki site (the Sphider
engine). This engine allows for cross-library searches and gives more
search results.  Also, the Doxygen search engine would sometimes give
weird results (dead links).

The output of the script that generates the library pages, can no longer
be viewed locally, because it now consists of PHP scripts (not plain
HTML) that have to be viewed on a webserver running PHP and require the
common/mcrl2.php file.

I intend to add an option to the script that allows for offline viewing.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@4330 21252fe5-60e8-0310-8ff0-c1ae1001594e