Improved performance of the pretty print functionality: + in function gsGroupDeclsBySort, replaced 2 quadratic loops by 3 linear ones + in function PRINT_FUNC(PrintDecls), replaced a quadratic loop by a linear one
Note that although the performance of the above functions increased
dramatically, this will rarely be noticable in practice, because the
size of the loops is the number of (consecutively) declared variables
(of data equations, summations, linear processes, etc), which is 2 in
most cases. However, I have seen a case (provided by Jan Friso Groote)
in which an LPS contained a declaration of 300.000 variables.


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@4346 21252fe5-60e8-0310-8ff0-c1ae1001594e