Added a --rewriter option to the lineariser. E.g. mcrl22lps -rinnerp ... now linearises using the prover.
Replaced the -r option as a shorthand for no-alpha by 
the -z option.

I failed to use the documentation of the command line
option to create the code. Only by inspection of lps2lts
I found out that I needed to use:

  options.rewrite_strategy        = parser.option_argument_as< RewriteStrategy >("rewriter");

In particular I have no idea how I can find out from the 
command line options that I should use the string "rewriter".
Either I grossly overlooked this, or the documentation needs
to be improved in the respect.

This also holds for invoking prover options.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@4718 21252fe5-60e8-0310-8ff0-c1ae1001594e