In the previous commit, when passing the -h/--help option to LTSView, the main window would be shown very briefly, before printing the help information to the command line and exiting the application.
This has now been fixed. The command line is first parsed and checked
for errors, after which the window is shown and any error messages
displayed. By parsing the command line first, the help message is
printed and the tool exited before the main window is drawn, when
passing the -h/--help option.

(Reviewed by Frank.)



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@4758 21252fe5-60e8-0310-8ff0-c1ae1001594e