Fixed a bug noted by Aad: the OK and Cancel buttons would sometimes be invisible on the Export to bitmap/vector dialogs, which is very annoying.
The problem was that the text label showing the directory path would
become too large, causing the buttons to move outside of the dialog.

The path to the selected file is now shown in a single text label. The
text string is truncated in the middle if it is longer than 50
characters.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@5003 21252fe5-60e8-0310-8ff0-c1ae1001594e