Added a workaround for a bug in Doxygen 1.5.7.1.
As noted by Frank Stappers, all pages would contain incorrect links to
pages that belong to other libraries (for example when linking to a page
in the ATerm++ library from a page in the LPS library). The links would
have the suffix ".html.php" which should be ".php".

This is now fixed by removing all occurrences of ".html" from every
tag-file that Doxygen generates. I will file a bug report to Doxygen if
this issue has not been reported yet.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@5006 21252fe5-60e8-0310-8ff0-c1ae1001594e