A solution for problem #520 is added by removing the deprecated function domain_sorts.
A solution for problem #518 PBES library cannot handle free function variables
is checked in. In case an expression of a given function sort is
required, a mapping of that sort is now also found.
Given that our language of sort expressions is now quit rich,
the function default_expression does not provide terms for a
given sort in all conceivable situations, and can be much
strengthened.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@5214 21252fe5-60e8-0310-8ff0-c1ae1001594e