Added a possibility to declare functions with the following names:
max  
min  
abs 
succ
pred
div
mod
exp
head
tail
rhead
rtail
in
count

with types that are different from the built-in types. 
For example it is possible to add function succ:Nat->Nat, 
but not possible to add function succ:Nat->Pos.
It is not possible to add constants with these names.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@5385 21252fe5-60e8-0310-8ff0-c1ae1001594e