A new version of lpsrealelm is now being committed.
It now works well on the ballgame for < 15 seconds, except that
1) due to the structure of the result of lpsuntime, it is currently better to
   untime by hand.
2) for larger values than appr. 15 seconds, application to the ballgame crashes, most
   likely due to unprotected aterms.

The biggest improvement is that the fourier-motzkin routine now
delivers a minimal set of equations. E.g. instead of [x<1, x<2] only
[x<1] is delivered. If the result is inconsistent, the result of
fourier-motzkin is [false].

For those not knowing the ballgame, it is the following. A 
player must repeatedly throw a ball for n seconds. If he does
not throw a ball within a second after the last throw, he is
disqualified. The resulting untimed state space for the ball
game is quite nice.





git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@5464 21252fe5-60e8-0310-8ff0-c1ae1001594e