The lineariser contained an optimisation to merge summands with the same action and nextstate parameters (modulo a renaming of variables). The conditions c1 and c2 of both summands were written as c1 || c2 in the new summand.
As on the one hand this could be an extremely costly 
operation (esp. if the input was already linear) and
also due to comments that the lineariser should not
meddle too much with its input, this optimisation has
been commented out, awaiting results on esp. the benchmark
test.

This optimisation was added in the first version of 
the lineariser, because it seemed to be a good idea 
back then.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@5717 21252fe5-60e8-0310-8ff0-c1ae1001594e