Added a forgotten case for assignments to the lineariser.
Adapted the test cases of the linearizer to match the interpretation
that variables in assignments refer back by name. So in

P(x:Nat)=sum x:Nat.a.P();

the x in the P assigned to x in the right hand side is the x
of the sum, and not the x in the left hand side. More precisely
the line above must be read as:

P(x:Nat)=sum x':Nat.a.P(x');

and not as:

P(x:Nat)=sum x':Nat.a.P(x);

Note that this means that with alpha conversion, it is needed
to sometimes change invisible objects. I.e. 

   sum x:Nat.P() equals sum x':Nat.P(x=x').

Jan Friso



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@5810 21252fe5-60e8-0310-8ff0-c1ae1001594e