Changed the communication function such that it will employ the values in communicating actions when calculating the result.
Improved the use of constant elimination in processes. The case where variables were replaced
by other variables, which were not adapted was not treated correctly. 

Found a bug in alpha conversion (due to an example by Jeroen Ketema). The bug was caused by
translating the old linearizer to the new linearizer using the new data library

Added several testcases covering the problematic situations.

Added several equations to the Othello (or Reversi) game to allow lpsparunfold and lpsconstelm
to reduce the game to elementary operations on the board. Due to persisting problems in lpsparunfold
this does not yet work.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@6918 21252fe5-60e8-0310-8ff0-c1ae1001594e