The bisimulation reduction algorithm has completely been rewritten. A new library containing a bisimulation partitioner has been added for divergence preserving branching, branching and strong bisimulation. These routines expect a transition system without tau loops.
A new partitioner class has been written to remove tau cycles and
leave them as tau loops on single states (for divergence preservation)
or remove them completely.

Divergence preserving branching bisimulation has been added to tools.

On the level of lts's it is now possible to call routines to reduce
the lts modulo above mentioned forms of bisimulation, or remove
connected tau components.

There is a testcase added to test the bisimulation reductions.

The code in bsimdata.cpp and bsimkernel.cpp has been removed almost
completely.

The algorithms have thoroughly been tested, but especially in the
connection to the tools (with large number of options) not everything
may have been connected well.

The performance of the algorithms is several factors better than the
old performance. The memory consumption for bisimulation reduction has
been reduced by a factor 2. There does not seem to be an explosion
in required memory anymore when writing the resulting lts to a file.

The bisimulation algorithm follows the Groote-Vaandrager algorithm
directly. There is room for improvement of this algorithm, by for
instance distributing states and transitions on a by need basis
over blocks, instead of distributing them immediately. Expectation is
that this will reduce running time by one or two factors.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@7103 21252fe5-60e8-0310-8ff0-c1ae1001594e