Reversed pbes2bool to use the old rewriter by Muck directly, instead of using the rewriter via the new interface. This should restore the performance. Changes are rather minimal, such that restoring the use of the new rewriter should be straightforward. The flags of pbes2bool have not been changed, but the use of -p.... (selecting a pbes rewriter) has currently no effect.
Added a method to the new rewriter interface to obtain the underlying rewriter.
This is not really elegant. I should consult with Wieger whether there is a
better method to do this.

Added a number of additional testcases for pbesses, especially a few where
quantifiers play a role.

All test cases pass in debug mode. I will check the performance and tests 
on the Olifant in non debug mode (for which this code first has to be checked in). There is a distinct
difference in the code being run in debug mode (which does not allow precompilation
of the pbes to rewrite format) and non debug mode (which by default uses precompilation).




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@7264 21252fe5-60e8-0310-8ff0-c1ae1001594e