In constelm it was possible that not all parameters of a process were set to correct values, because process_initializer.state() does not give all parameter values for a process, but only those that change. So, given the process
  P(x:Nat)= ....

with initial value 

  P(x) 

where x is a variable of sort Nat, process_initializer.state()
would be empty, and in constelm a substitution would be constructed
mapping x to a random place in memory. This problem only occurs when
manipulating intermediate linear processes (which happens in the linearizer)
because in general all parameters of a linear process are set to closed
data expressions, and hence the situation above cannot occur. 

This problem is repaired by constructing the substitution out of
the initializer, without looking at the parameters.

It is necessary that process_initializer.state() is investigated
because it yields a confusing result.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@7449 21252fe5-60e8-0310-8ff0-c1ae1001594e