Changed the normalisation of function sorts. Up till now, if a sort declaration of the form
sort F=A->B;

was declared, every occurrence of a type A->B was
replaced by F when normalising sorts. However, at many
places in the toolset, it is checked whether a sort 
is a function sort, and whether the arity of the function
type is correct. In the case above the sort F is not
a function sort, and has no arity, contrary to the
A->B which is a function sort.

By normalising each alias with a function sort from left
to right this problem is avoided.

Aliases with container sorts and structured sorts are
still normalised from right to left. For structured sorts
this is essential, because these can be recursive. For
container sorts either choice can be made.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@7697 21252fe5-60e8-0310-8ff0-c1ae1001594e