Restructured the lts class for labelled transition systems. Moved algorithms out of the class (e.g. the bisimulation reduction algorithm should not be a member of the lts class). All algorithms are moved to a new header file lts_algorithm.h.
Removed the class p_lts. Added explanation on how to use
reductions in quickbook. Started with an explanation of the
partitioner algorithms. 

The lts class is still quite awkward, and further restructuring
is in order. As it is unlikely that this will happen in one big
sweep, it appears to be a good strategy to do this in a number of
smaller steps.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@7737 21252fe5-60e8-0310-8ff0-c1ae1001594e