- Modified next_state_generator class such that it can actually be used for state space generation. - Added test for state space generation of ABP.
It turns out that the iterator interface behaves slightly different from normal. Only after doing an increment it can be determined whether a new state has been found. So it now looks like this:

  next_state_generator generator(lps_spec);
  next_state_generator::iterator first = generator.begin();
  next_state_generator::iterator last;
  while (++first != last)
  {
    std::cout << generator.print_state(*first) << std::endl;
  }

Since the behavior of the iterators is non-standard it might be better to change it into:

  next_state_generator generator(lps_spec);
  next_state_generator::iterator i = generator.begin();
  while (++i)
  {
    std::cout << generator.print_state(*i) << std::endl;
  }


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@7747 21252fe5-60e8-0310-8ff0-c1ae1001594e