- Added the wrapper class mcrl2_gui_tool, that adds the option --mcrl2-gui to a tool. - Added default constructors to squadt_tool.
The goal of this exercise is to add support for Frank's squadt replacement tool (called mcrl2-gui). Another goal is to separate the squadt and mcrl2-gui from the command line tools. It turns out that separating squadt from the command line tools introduces a few problems. Apparently the function synchronise_with_configuration needs to be specialized for each tool, which is ugly. Also the graphical tools currently seem to depend on squadt.

A working example in which the command line tool is separated from both squadt and 'mcrl2-gui' has been made for the tool lpsinfo.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@7825 21252fe5-60e8-0310-8ff0-c1ae1001594e