- Added some search functions to lps/find.h. - Made some operators public in the binding aware traversers, to make it compile. - Added a simple test case for search_free_variable. Unfortunately it doesn't compile with Visual C++.
The search_free_variable function uses the class selective_binding_aware_traverser, which is not fully understood yet. It is also unknown whether or not this approach can be made to work with Visual C++.

Currently in linearise.cpp the function data::search_free_variable is applied to an LPS data type, which is incorrect. It accidentally works with g++ through a deprecated aterm overload, and it does not work with Visual C++. The intended solution is to use lps::search_free_variable instead, but at this moment it is unknown how to implement it.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@7857 21252fe5-60e8-0310-8ff0-c1ae1001594e