The function ATconcat AT_alloc_protected to reserve some workspace. As this last function has a lousy garbage collection mechanism, it is very slow, at times.
It is much faster to allocate the memory required on the stack.

This is done now, but it might be that some compilers cannot
deal with this properly. I currently check this in. There are
quite a number of other spots where AT_alloc_protected is called
to reserve temporary memory, and if the current check in passes
all the tests, I am in favour of removing all unnessecary calls
to AT_alloc_protected in the ATerm library.

My expectation is that as the ATerm library is not really
maintained anymore, we are more or less forced to maintain it
ourselves, and can make such changes as well.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8033 21252fe5-60e8-0310-8ff0-c1ae1001594e