After discussion with Muck regarding reserving memory in the list.c from the ATerm library, I decided to use an old construct which tests whether MSVC is used, and in that case uses _alloca, and otherwise declares a dynamic array on the stack.
We are not sure whether MSVC is recognized correctly,
but I have no machine to directly test this on.

In certain cases this change gave a 10-fold
speed improvement in various tools. The problems
were often caused by some piece of code messing
up the very naive garbage collection algorithm 
by Eric Scheffers, which would subsequently 
cause quite a number of list operations to spend
lots of time in searching in the garbage collection
data structure. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8036 21252fe5-60e8-0310-8ff0-c1ae1001594e