- Added a test case for which the monotonicity check fails, based on an example supplied by Jeroen Keiren. It has been reduced to the formula !(mu X. (mu X. X)). The reason that is fails is the nesting of the same variable X. The algorithm that checks for monotonicity does not take such name clashes into account.
This case will be solved by adding a preprocessing step that resolves name clashes. Also the translation of a regular formula to a state formula will be replaced. First of all it is outdated ATerm code that is dependent on the internal format, and second this translation introduces such name clashes.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8224 21252fe5-60e8-0310-8ff0-c1ae1001594e