- Put class selective_traverser in a separate header file. - Generated state_formula classes and visitors. - Added a checker for name clashes in state formulas based on traversers (it is still unused at the moment). - Removed the state_frm namespace, and updated the code that relied on it.
There is one unsolved problem. The data library uses classes to represent expressions that derive from data_expression. This idea has been copied to other libraries. But it turns out there is a nasty problem with unary function expressions. Take for example the class not_ that derives from state_formula. In some contexts we want not_(x) to be the application of the negation to x, while in other contexts we want it to be construction of a not_ with the term x, without applying negation. This leads to subtle mistakes, where silently the wrong choice is taken. In some cases it also introduces ambiguities when compiling.


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8226 21252fe5-60e8-0310-8ff0-c1ae1001594e