It turned out that the philosophy that sort aliases were rewritten from right to left led to the problem that it was not possible to recognize the nature of a sort based on the sort expression. For instance, the sort alias
sort F = G->H

would cause any sort G->H to be replaced by F. F is a basic
sort, and not a function sort, so F would mistakenly be interpreted
as a NON function sort.

Although the means were available to reconstruct G->H from F the 
provided functions were nowhere used in the toolset. Therefore, 
I decided to revert the direction in which sorts are normalised.
All sorts aliases are now rewritten from left to right, except
structured sort declarations, which are interpreted from right
to left. This last step is necessary, because structured sorts
can be recursive.

All tests succeed. A number of pbes tests are failing but this
is not related to the changes in the way aliases are dealt with.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8235 21252fe5-60e8-0310-8ff0-c1ae1001594e