Reimplemented the .lts formats. The purpose is to reduce the complexity of the lts library. The basic idea is that there is one lts_ format, which has action labels and state labels as a template parameter. All the known formats are template instantiations of the lts format. The current formats are
lts 
aut
fsm
bcg 
dot
svc

The lts format has timed multi actions as transition labels,
and vectors of data expression as state labels. 
The aut format has no state labels, and strings as transition
labels. 
The fsm format very much resembles the .fsm format, by
having vectors of numbers as state labels (where the numbers
refer to vectors in a separate array) and strings as transition
labels. etc.

All algorithms are templated, and work on all formats.
There is a (unfinished) set of translations between the
formats. Every translation is quite straightforward. But
with 6 formats, there are 36 required translations.

The graphical tools now work with the (internal) fsm
format. This means that they are not dependent on the
ATerm library anymore, but use strings instead.

A small problem with the data normalisation has also
been addressed. A small part was not implemented, but
it was required when structured types were used without
explicit definition, eg. sort L=List(struct a|b);

All tests pass, but most likely errors will still
pop up to be dealt with in the coming days.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8307 21252fe5-60e8-0310-8ff0-c1ae1001594e