Work around class between doxygen and mcrl2 css
The issue that is worked around is the following. The mCRL2 homepage uses the CSS class "header" for the webpage header (with gradient). The new HTML/PHP generated by Doxygen (as of Doxygen 1.7.0) also uses the CSS class "header". This conflicts, and messes up the library reference documentation.

For now, we work around this solution by adding an appropriate definition for the class "doxyheader" in doxystyle.css, and replacing all occurrences of class="header" by class="doxyheader" in the doxygen generated output (using sed).

TODO:
- Find an appropriate solution for this issue, e.g. by prefixing all mcrl2 homepage classes with and mcrl2 prefix. This could cause issues with the wiki page though.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8380 21252fe5-60e8-0310-8ff0-c1ae1001594e