Fix exit status of typecheck test.
It turns out that the exit status 200 was caused by *unexpected* errors, that were present next to expected errors. Now all errors are treated as expected, and the exit status of the test is 0 (no errors).

Note that I have slightly modified some checks in the data library for improved usability.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8407 21252fe5-60e8-0310-8ff0-c1ae1001594e