Took care that the no-constelm programme option shows up in mcrl2-gui.
Repaired a bug where (for whatever reason) variables in a where clause
had a wrong shape. The problem was that the typechecker assumed that a 
variable was not yet typed, and therefore transformed it into 
DataVarId(DataVarId("z",SortId("Bool")),SortId("Bool")). With an
extra check this is now avoided.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8416 21252fe5-60e8-0310-8ff0-c1ae1001594e