Split data_expression_with variables.h
This commit essentially splits the file in the class definition, and a file that defines the expression traits. This splits the cyclic dependencies that prevented the class from being included in traverser.h. The complete code now compiles with clang. Compiled and tested with clang and gcc.

Test results:
clang: all tests, excepts lps_monotonicity_test pass. This test failure is similar to the one with the intel build, and should be investigated. Note: looks like a garbage collection issue.
g++: all tests pass, as expected.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8441 21252fe5-60e8-0310-8ff0-c1ae1001594e